+++ test/rules/one_statement_per_block_test.dart
@@ -9,7 +9,7 @@ void main() {
 }
 
 @reflectiveTest
-class OneStatementPerBlockTest extends AnalysisRuleTest {
+final class OneStatementPerBlockTest extends AnalysisRuleTest {
   @override
   void setUp() {
     rule = OneStatementPerBlockLint();